don't call worker_exit from the Arbiter
When workers are killed by the Arbiter, don't call worker_exit. It's
confusing to have worker_exit called from the Arbiter only when the
worker is killed explicitly. When workers exit normally and are reaped
after SIG_CHLD the hook is called from the worker process context.

Calling the same handler from both contexts is difficult because the
state of the worker has diverged at that point. There should be no use
for users to see the worker in the pre-fork state (Arbiter perspective)
here.